content/media/video/test/test_autobuffer2.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=479863
-->
<head>
  <title>Test for Bug 479863</title>
  <script type="application/javascript" src="/MochiKit/packed.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript" src="use_large_cache.js"></script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=479863">Mozilla Bug 479863</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>

<script type="application/javascript">
var stalls = { v1:false, v2:false, v3:false, v4:false, v5:false, v6:false };
var loads = { v1:false, v2:false, v3:false, v4:false, v5:false, v6:false };
var totalLoadEvents = 0;
var expectedLoadEvents = 6;

function loaded(event) {
  loads[event.target.id] = true;
  ++totalLoadEvents;
  if (totalLoadEvents == expectedLoadEvents) {
    ok(stalls.v1, "v1 should stall");
    // We might get accidental stalls on very slow machines, so lets not test these
    // ok(!stalls.v2, "v2 should not stall");
    // ok(!stalls.v3, "v3 should not stall");
    ok(stalls.v4, "v4 should stall");
    ok(stalls.v5, "v5 should stall");
    ok(stalls.v6, "v6 should stall");
    ok(loads.v1, "v1 should load after setting autobuffer");
    ok(loads.v2, "v2 should load");
    ok(loads.v3, "v3 should load");
    ok(loads.v4, "v4 should load after calling play()");
    ok(loads.v5, "v5 should load after seeking");
    ok(loads.v6, "v6 should load after setting autoplay");
    SimpleTest.finish();
  }
}
function stalled(event) {
  stalls[event.target.id] = true;
  if (event.target.id == "v1") {
    event.target.autobuffer = true;
  } else if (event.target.id == "v4") {
    event.target.play();
  } else if (event.target.id == "v5") {
    event.target.currentTime = 0.1;
  } else if (event.target.id == "v6") {
    event.target.autoplay = true;
  }
}
</script>

<video id='v1' src="seek.ogv" onstalled="stalled(event)" onload="loaded(event)"></video>
<video id='v2' src="seek.ogv" onstalled="stalled(event)" onload="loaded(event)" autobuffer></video>
<video id='v3' src="seek.ogv" onstalled="stalled(event)" onload="loaded(event)" autoplay></video>
<video id='v4' src="seek.ogv" onstalled="stalled(event)" onload="loaded(event)"></video>
<video id='v5' src="seek.ogv" onstalled="stalled(event)" onload="loaded(event)"></video>
<video id='v6' src="seek.ogv" onstalled="stalled(event)" onload="loaded(event)"></video>

<pre id="test">
<script type="application/javascript">
SimpleTest.waitForExplicitFinish();
</script>
</pre>
</body>
</html>
